auto: enhanced_core/parallel_executor.py - def execute_parallel(self,

+++ enhanced_core/parallel_executor.py
@@ -48,7 +48,7 @@ class ParallelExecutor:
 
     def execute_parallel(self,
                          tasks: Dict[str, Callable],
-                         timeout: Optional[int] = 30) -> Dict[str, TaskResult]:
+                         timeout: Optional[float] = 30.0) -> Dict[str, TaskResult]:
         """
         并行执行多个任务
